moved testdata to test/testtools.ts to make sure importing testtools.ts does not define tests
